requirements.txt
flask==2.3.3
gunicorn==21.2.0
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.0
Werkzeug==2.3.7